[java] break out enums and remove reflection
